edge/modules/factory-simulator/requirements.txt
azure-iot-device==2.12.0
numpy==1.24.3
scipy==1.11.1
loguru==0.7.2
pydantic==2.4.2
python-dotenv==1.0.0
aiofiles==23.2.0
schedule==1.2.0